+++ package.json
@@ -26,12 +26,12 @@
     "leaflet": "1.5.1",
     "leaflet-hotline": "0.4.0",
     "local-storage-fallback": "4.1.1",
-    "pica": "5.0.0",
+    "pica": "5.1.0",
     "prop-types": "15.7.2",
     "query-string": "6.8.1",
     "react": "16.8.6",
     "react-bootstrap": "0.32.4",
-    "react-bootstrap-typeahead": "3.4.5",
+    "react-bootstrap-typeahead": "3.4.6",
     "react-chartjs-2": "2.7.6",
     "react-dom": "16.8.6",
     "react-dropzone": "10.1.5",
@@ -42,7 +42,7 @@
     "react-redux": "7.1.0",
     "react-stars": "2.2.5",
     "react-tag-autocomplete": "5.11.1",
-    "redux": "4.0.1",
+    "redux": "4.0.4",
     "redux-logger": "3.0.6",
     "redux-logic": "2.1.1",
     "rxjs": "6.5.2",
@@ -50,53 +50,53 @@
     "typescript-is": "0.12.2"
   },
   "devDependencies": {
-    "@babel/core": "7.4.5",
+    "@babel/core": "7.5.5",
     "@babel/polyfill": "7.4.4",
-    "@babel/preset-env": "7.4.5",
-    "@types/chart.js": "2.7.54",
+    "@babel/preset-env": "7.5.5",
+    "@types/chart.js": "2.7.55",
     "@types/google.analytics": "0.0.40",
     "@types/history": "4.7.2",
-    "@types/react": "16.8.22",
-    "@types/react-bootstrap": "0.32.18",
+    "@types/react": "16.8.23",
+    "@types/react-bootstrap": "0.32.19",
     "@types/react-bootstrap-typeahead": "3.4.5",
     "@types/react-dom": "16.8.4",
     "@types/react-leaflet": "2.2.1",
-    "@types/react-rangeslider": "2.2.0",
+    "@types/react-rangeslider": "2.2.1",
     "@types/react-redux": "7.1.1",
     "@types/react-tag-autocomplete": "5.6.1",
     "babel-eslint": "10.0.2",
     "babel-loader": "8.0.6",
     "copy-webpack-plugin": "5.0.3",
-    "css-loader": "3.0.0",
+    "css-loader": "3.1.0",
     "cssnano": "4.1.10",
     "ejs-loader": "0.3.3",
     "eslint": "6.0.1",
     "eslint-config-prettier": "6.0.0",
-    "eslint-loader": "2.1.2",
+    "eslint-loader": "2.2.1",
     "eslint-plugin-prettier": "3.1.0",
     "eslint-plugin-react": "7.14.2",
     "eslint-plugin-react-hooks": "1.6.1",
-    "file-loader": "4.0.0",
+    "file-loader": "4.1.0",
     "html-loader": "0.5.5",
     "html-webpack-plugin": "3.2.0",
     "markdown-loader": "5.0.0",
-    "mini-css-extract-plugin": "0.7.0",
+    "mini-css-extract-plugin": "0.8.0",
     "node-sass": "4.12.0",
     "offline-plugin": "5.0.7",
     "optimize-css-assets-webpack-plugin": "5.0.3",
     "prettier": "1.18.2",
-    "raw-loader": "3.0.0",
+    "raw-loader": "3.1.0",
     "sass-loader": "7.1.0",
     "source-map-loader": "0.2.4",
     "style-loader": "0.23.1",
     "ts-loader": "6.0.4",
     "ttypescript": "1.5.7",
-    "typescript": "3.5.2",
+    "typescript": "3.5.3",
     "uglifyjs-webpack-plugin": "2.1.3",
-    "url-loader": "2.0.1",
-    "webpack": "4.35.0",
+    "url-loader": "2.1.0",
+    "webpack": "4.36.1",
     "webpack-cleanup-plugin": "0.5.1",
-    "webpack-cli": "3.3.5",
+    "webpack-cli": "3.3.6",
     "webpack-dev-server": "3.7.2",
     "webpack-pwa-manifest": "4.0.0"
   },
